+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Add `iosxr_extcommunity_rt_set` resource and data source
 - Add `iosxr_extcommunity_soo_set` resource and data source
 - Add `iosxr_fpd` resource and data source
+- Add `iosxr_extcommunity_cost_set` resource and data source
+- Add `iosxr_rd_set` resource and data source
 
 ## 0.2.1